Parameterize path and filename of token file
Added optional parameter which, when provided by the object instantiator, results in tokens to be stored at the given location and filename. This allows, for example, storing the tokens in /tmp to keep the tokens from being written to the local drive. The author stores the token file on a secure server and copies it to /tmp as needed.

The change is backward compatible: If the parameter is not provided, the prior behavior of storing the tokens in ts_state.json in the working directory is maintained.